Refactored sidebar example, fix summary charts crash in sidebar example
